Hook to reactor before starting it
Doing a listenTCP on the reactor after it is started was flaky at best. It did not work on OS X
and was not best practice. The port would be in the listening state, but the reactor would never
be aware of connections made to the port.

This provides a mechanism for tests to hook onto the reactor being prepared, right before
launching it.