Add best-effort reset_interface on dell
Add a reset_interface on dell which deletes the attributes it supports
and clean the interface to the most of his knowledge, this allows to
strip the mtu from the interface on reset.

Fake-switches 1.1.6 now features the command 'no switchport mode' that
allows to set the interface in access (default) mode and remove
mode-related attributes.