Clean ssh connection in destructor
Netman was leaking file descriptor with
established connection when runninng unit test.
It got up to a 1000 connections running and was
problematic because on osx the limit is now 256 by
default.
To reproduce, use a commit prior this one and run
in a shell with ulimit -n 256.

This also lower cpu usage during unit testing for some
reason.